requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apache-beam==2.49.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes @ https://github.com/jllllll/bitsandbytes-windows-webui/raw/main/bitsandbytes-0.38.1-py3-none-any.whl#sha256=26e4f9590836c9a36cad5f0e7235ddf32639529b211fa8946edf99c9d4d4f497
black==23.7.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
botbuilder-core==4.14.4
botbuilder-schema==4.14.4
botframework-connector==4.14.4
botframework-streaming==4.14.4
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@7d2e68dbe452fc422268d40ac185ea2609affca8
dill==0.3.7
dnspython==2.4.1
docopt==0.6.2
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.4.0
elasticsearch==8.9.0
emoji==2.8.0
et-xmlfile==1.1.0
executing==1.2.0
fairscale==0.4.13
faiss-cpu==1.7.4
fastapi==0.99.1
fastavro==1.8.2
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.4
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
hdfs==2.7.1
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath @ git+https://github.com/facebookresearch/iopath@8f359cf63bec091969ad8f30fa14c17d3218c347
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isodate==0.6.1
isoduration==20.11.0
jax==0.4.14
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpickle==1.4.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.12.0
kiwisolver==1.4.4
langchain==0.0.270
langcodes==3.3.0
langsmith==0.0.25
layoutparser==0.3.4
libclang==16.0.6
# Editable Git install with no remote (llama==0.0.1)
-e c:\ai_chatbot\llama_2
loguru==0.7.0
lxml==4.9.3
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msal==1.23.0
msg-parser==1.2.0
msrest==0.6.21
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
objsize==0.6.1
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.8
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathy==0.10.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.10.0
portalocker==2.7.0
posthog==3.0.2
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.1
pydeck==0.8.0
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.4.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.15.4
PyPDF2==3.0.1
pypdfium2==4.18.0
PyPika==0.48.9
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
sqlitedict==2.1.0
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.14.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tornado==6.3.2
tqdm==4.65.2
traitlets==5.9.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.4
unstructured-inference==0.5.14
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.8
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
